refactor: `WebConfig`가 하나의 역할만 하도록 수정
