+++ backend/src/main/java/reviewme/config/cors/CorsConfig.java
@@ -1,4 +1,4 @@
-package reviewme.config;
+package reviewme.config.cors;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
