+++ backend/src/main/java/reviewme/config/datasource/DataSourceType.java
@@ -1,4 +1,4 @@
-package reviewme.config;
+package reviewme.config.datasource;
 
 public enum DataSourceType {
     READ,
