+++ backend/src/main/java/reviewme/config/datasource/ReplicationRoutingDataSource.java
@@ -1,4 +1,4 @@
-package reviewme.config;
+package reviewme.config.datasource;
 
 import org.springframework.jdbc.datasource.lookup.AbstractRoutingDataSource;
 import org.springframework.transaction.support.TransactionSynchronizationManager;
