+++ backend/src/main/java/reviewme/config/requestlimit/RequestLimitProperties.java
@@ -1,13 +1,8 @@
-package reviewme.config;
+package reviewme.config.requestlimit;
 
 import java.time.Duration;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 
 @ConfigurationProperties(prefix = "request-limit")
-public record RequestLimitProperties(
-        long threshold,
-        Duration duration,
-        String host,
-        int port
-) {
+public record RequestLimitProperties(long threshold, Duration duration, String host, int port) {
 }
